+++ R/ggplot_box_fun.R
@@ -115,6 +115,7 @@ theme_box <-
 #' @param x_var Unquoted categorical variable to be on the x axis. Required input.
 #' @param y_var Unquoted numeric variable to be on the y axis. Defaults to NULL. Required if stat equals "boxplot".
 #' @param stat String of "boxplot" or "identity". Defaults to "boxplot". If identity is selected, data provided must be grouped by the x_var with ymin, lower, middle, upper, ymax variables. Note "identity" does not provide outliers.
+#' @param x_scale_labels Argument to adjust the format of the x scale labels.
 #' @param y_scale_zero TRUE or FALSE of whether the minimum of the y scale is zero. Defaults to TRUE.
 #' @param y_scale_trans TRUEransformation of y-axis scale (e.g. "signed_sqrt"). Defaults to "identity", which has no transformation.
 #' @param y_scale_labels Argument to adjust the format of the y scale labels.
@@ -160,6 +161,7 @@ ggplot_box <- function(data,
                        x_var,
                        y_var = NULL,
                        stat = "boxplot",
+                       x_scale_labels = waiver(),
                        y_scale_zero = TRUE,
                        y_scale_trans = "identity",
                        y_scale_labels = waiver(),
@@ -255,6 +257,7 @@ ggplot_box <- function(data,
   }
   
   plot <- plot +
+    scale_x_discrete(labels = x_scale_labels) +
     scale_y_continuous(
       expand = c(0, 0),
       breaks = y_scale_breaks,
@@ -272,11 +275,7 @@ ggplot_box <- function(data,
         x = stringr::str_wrap(x_title, wrap_x_title),
         y = stringr::str_wrap(y_title, wrap_y_title),
         caption = stringr::str_wrap(caption, wrap_caption)
-      ) +
-      scale_x_discrete(
-        labels = function(x)
-          stringr::str_wrap(x, 15)
-      )
+      ) 
   }
   else if (isMobile == TRUE){
     plot <- plot +
@@ -288,13 +287,8 @@ ggplot_box <- function(data,
         caption = stringr::str_wrap(caption, 50)
       ) +
       coord_flip() +
-      scale_x_discrete(
-        labels = function(x)
-          stringr::str_wrap(x, 30)
-      ) +
       theme(panel.grid.major.x = element_line(colour = "#D3D3D3", size = 0.2)) +
       theme(panel.grid.major.y = element_blank())
-    
   }
   
   return(plot)
@@ -307,6 +301,7 @@ ggplot_box <- function(data,
 #' @param y_var Unquoted numeric variable to be on the y axis. Defaults to NULL. Required if stat equals "boxplot".
 #' @param facet_var Unquoted categorical variable to facet the data by. Required input.
 #' @param stat String of "boxplot" or "identity". Defaults to "boxplot". If identity is selected, data provided must be grouped by the x_var and facet_var with ymin, lower, middle, upper, ymax variables. Note "identity" does not provide outliers.
+#' @param x_scale_labels Argument to adjust the format of the x scale labels.
 #' @param y_scale_zero TRUE or FALSE of whether the minimum of the y scale is zero. Defaults to TRUE.
 #' @param y_scale_trans TRUEransformation of y-axis scale (e.g. "signed_sqrt"). Defaults to "identity", which has no transformation.
 #' @param y_scale_labels Argument to adjust the format of the y scale labels.
@@ -347,6 +342,7 @@ ggplot_box_facet <-
            y_var = NULL,
            facet_var,
            stat = "boxplot",
+           x_scale_labels = waiver(),
            y_scale_zero = TRUE,
            y_scale_trans = "identity",
            y_scale_labels = waiver(),
@@ -458,6 +454,9 @@ ggplot_box_facet <-
                            oob = scales::rescale_none)
     }
     
+    plot <- plot +
+      scale_x_discrete(labels = x_scale_labels)
+      
     if (isMobile == FALSE){
       if (is.null(facet_nrow) & length(unique(facet_var_vector)) <= 3) facet_nrow <- 1
       if (is.null(facet_nrow) & length(unique(facet_var_vector)) > 3) facet_nrow <- 2
@@ -469,10 +468,6 @@ ggplot_box_facet <-
           x = stringr::str_wrap(x_title, wrap_x_title),
           y = stringr::str_wrap(y_title, wrap_y_title),
           caption = stringr::str_wrap(caption, wrap_caption)
-        ) +
-        scale_x_discrete(
-          labels = function(x)
-            stringr::str_wrap(stringr::str_replace_all(x, "__.+$", ""), 15)
         ) +
         facet_wrap(vars(!!facet_var), scales = facet_scales, nrow = facet_nrow)
     }
@@ -487,13 +482,8 @@ ggplot_box_facet <-
         ) +
         facet_wrap(vars(!!facet_var), scales = facet_scales, ncol = 1) +
         coord_flip() +
-        scale_x_discrete(
-          labels = function(x)
-            stringr::str_wrap(stringr::str_replace_all(x, "__.+$", ""), 30)
-        ) +
         theme(panel.grid.major.x = element_line(colour = "#D3D3D3", size = 0.2)) +
         theme(panel.grid.major.y = element_blank())
-      
     }
     
     return(plot)