+++ R/ggplot_vbar_fun.R
@@ -114,7 +114,7 @@ theme_vbar <-
 #' @param x_var Unquoted numeric, date or categorical variable to be on the x axis. Required input.
 #' @param y_var Unquoted numeric variable to be on the y axis. Required input.
 #' @param hover_var Unquoted variable to be an additional hover variable for when used inside plotly::ggplotly(). Defaults to NULL.
-#' @param x_scale_date_format Date format for x axis labels.
+#' @param x_scale_labels Argument to adjust the format of the x scale labels.
 #' @param y_scale_zero TRUE or FALSE of whether the minimum of the y scale is zero. Defaults to TRUE.
 #' @param y_scale_trans A string specifying a transformation for the y axis scale, such as "log10" or "sqrt". Defaults to "identity".
 #' @param y_scale_labels Argument to adjust the format of the y scale labels.
@@ -155,7 +155,7 @@ ggplot_vbar <- function(data,
                         x_var,
                         y_var,
                         hover_var = NULL,
-                        x_scale_date_format = "%Y",
+                        x_scale_labels = waiver(),
                         y_scale_zero = TRUE,
                         y_scale_trans = "identity",
                         y_scale_labels = waiver(),
@@ -284,7 +284,7 @@ ggplot_vbar <- function(data,
       scale_x_date(
         expand = c(0, 0),
         breaks = x_scale_breaks,
-        labels = scales::date_format(x_scale_date_format)
+        labels = x_scale_labels
       )
   }
   else if (is.numeric(x_var_vector)) {
@@ -296,11 +296,12 @@ ggplot_vbar <- function(data,
     plot <- plot +
       scale_x_continuous(expand = c(0, 0),
                          breaks = x_scale_breaks,
+                         labels = x_scale_labels,
                          oob = scales::rescale_none)
   }
   else if (is.character(x_var_vector) | is.factor(x_var_vector)){
     plot <- plot +
-      scale_x_discrete(expand = c(0, 0))
+      scale_x_discrete(expand = c(0, 0), labels = x_scale_labels)
   }
 
   plot <- plot +
@@ -371,7 +372,7 @@ ggplot_vbar <- function(data,
 #' @param y_var Unquoted numeric variable to be on the y axis. Required input.
 #' @param col_var Unquoted categorical variable to colour the bars. Required input.
 #' @param hover_var Unquoted variable to be an additional hover variable for when used inside plotly::ggplotly(). Defaults to NULL.
-#' @param x_scale_date_format Date format for x axis labels.
+#' @param x_scale_labels Argument to adjust the format of the x scale labels.
 #' @param y_scale_zero TRUE or FALSE of whether the minimum of the y scale is zero. Defaults to TRUE.
 #' @param y_scale_trans A string specifying a transformation for the y axis scale, such as "log10" or "sqrt". Defaults to "identity".
 #' @param y_scale_labels Argument to adjust the format of the y scale labels.
@@ -416,7 +417,7 @@ ggplot_vbar_col <-
            y_var,
            col_var,
            hover_var = NULL,
-           x_scale_date_format = "%Y",
+           x_scale_labels = waiver(),
            y_scale_zero = TRUE,
            y_scale_trans = "identity",
            y_scale_labels = waiver(),
@@ -584,7 +585,7 @@ ggplot_vbar_col <-
         scale_x_date(
           expand = c(0, 0),
           breaks = x_scale_breaks,
-          labels = scales::date_format(x_scale_date_format)
+          labels = x_scale_labels
         )
     }
     else if (is.numeric(x_var_vector)) {
@@ -596,11 +597,12 @@ ggplot_vbar_col <-
       plot <- plot +
         scale_x_continuous(expand = c(0, 0),
                            breaks = x_scale_breaks,
+                           labels = x_scale_labels,
                            oob = scales::rescale_none)
     }
     else if (is.character(x_var_vector) | is.factor(x_var_vector)){
       plot <- plot +
-        scale_x_discrete(expand = c(0, 0))
+        scale_x_discrete(expand = c(0, 0), labels = x_scale_labels)
     }
 
     plot <- plot +
@@ -667,7 +669,7 @@ ggplot_vbar_col <-
 #' @param y_var Unquoted numeric variable to be on the y axis. Required input.
 #' @param facet_var Unquoted categorical variable to facet the data by. Required input.
 #' @param hover_var Unquoted variable to be an additional hover variable for when used inside plotly::ggplotly(). Defaults to NULL.
-#' @param x_scale_date_format Date format for x axis labels.
+#' @param x_scale_labels Argument to adjust the format of the x scale labels.
 #' @param y_scale_zero TRUE or FALSE of whether the minimum of the y scale is zero. Defaults to TRUE.
 #' @param y_scale_trans A string specifying a transformation for the y axis scale, such as "log10" or "sqrt". Defaults to "identity".
 #' @param y_scale_labels Argument to adjust the format of the y scale labels.
@@ -711,7 +713,7 @@ ggplot_vbar_facet <-
            y_var,
            facet_var,
            hover_var = NULL,
-           x_scale_date_format = "%Y",
+           x_scale_labels = waiver(),
            y_scale_zero = TRUE,
            y_scale_trans = "identity",
            y_scale_labels = waiver(),
@@ -843,7 +845,7 @@ ggplot_vbar_facet <-
           scale_x_date(
             expand = c(0, 0),
             breaks = x_scale_breaks,
-            labels = scales::date_format(x_scale_date_format)
+            labels = x_scale_labels
           )
       }
       else if (is.numeric(x_var_vector)) {
@@ -855,11 +857,12 @@ ggplot_vbar_facet <-
         plot <- plot +
           scale_x_continuous(expand = c(0, 0),
                              breaks = x_scale_breaks,
+                             labels = x_scale_labels,
                              oob = scales::rescale_none)
       }
       else if (is.character(x_var_vector) | is.factor(x_var_vector)){
         plot <- plot +
-          scale_x_discrete(expand = c(0, 0))
+          scale_x_discrete(expand = c(0, 0), labels = x_scale_labels)
       }
 
     }
@@ -962,7 +965,7 @@ ggplot_vbar_facet <-
 #' @param col_var Unquoted categorical variable to colour the bars. Required input.
 #' @param facet_var Unquoted categorical variable to facet the data by. Required input.
 #' @param hover_var Unquoted variable to be an additional hover variable for when used inside plotly::ggplotly(). Defaults to NULL.
-#' @param x_scale_date_format Date format for x axis labels.
+#' @param x_scale_labels Argument to adjust the format of the x scale labels.
 #' @param y_scale_zero TRUE or FALSE of whether the minimum of the y scale is zero. Defaults to TRUE.
 #' @param y_scale_trans A string specifying a transformation for the y axis scale, such as "log10" or "sqrt". Defaults to "identity".
 #' @param y_scale_labels Argument to adjust the format of the y scale labels.
@@ -1015,7 +1018,7 @@ ggplot_vbar_col_facet <-
            col_var,
            facet_var,
            hover_var = NULL,
-           x_scale_date_format = "%Y",
+           x_scale_labels = waiver(),
            y_scale_zero = TRUE,
            y_scale_trans = "identity",
            y_scale_labels = waiver(),
@@ -1185,7 +1188,7 @@ ggplot_vbar_col_facet <-
           scale_x_date(
             expand = c(0, 0),
             breaks = x_scale_breaks,
-            labels = scales::date_format(x_scale_date_format)
+            labels = x_scale_labels
           )
       }
       else if (is.numeric(x_var_vector)) {
@@ -1197,11 +1200,12 @@ ggplot_vbar_col_facet <-
         plot <- plot +
           scale_x_continuous(expand = c(0, 0),
                              breaks = x_scale_breaks,
+                             labels = x_scale_labels,
                              oob = scales::rescale_none)
       }
       else if (is.character(x_var_vector) | is.factor(x_var_vector)){
         plot <- plot +
-          scale_x_discrete(expand = c(0, 0))
+          scale_x_discrete(expand = c(0, 0), labels = x_scale_labels)
       }
     }