+++ NEWS.md
@@ -1,3 +1,9 @@
+# simplevis 1.4.7
+
+* added x_scale_labels and y_scale_labels for all places where missing from plots.
+* removed x_scale_date_format from plots.
+* removed wrap_y_labels argument from plots.
+
 # simplevis 1.4.6
 
 * Update line functions with lines = FALSE argument.